Updates package.json and admission-policy.module.ts, adds invitation.module.ts, and modifies auth.controller.ts and auth.module.ts
