+++ backend/package.json
@@ -16,7 +16,7 @@
     "lint": "eslint \"{src,apps,libs,test}/**/*.ts\" --fix",
     "test": "jest",
     "db:up": "docker-compose up -d",
-    "db:down": "docker-compose down -v", 
+    "db:down": "docker-compose down -v",
     "dev:all": "npm run db:up && npm run start:dev",
     "test:watch": "jest --watch",
     "test:cov": "jest --coverage",
