+++ backend/src/admission-policy/admission-policy.module.ts
@@ -2,6 +2,6 @@ import { Module } from '@nestjs/common';
 import { AdmissionPolicyService } from './admission-policy.service';
 
 @Module({
-  providers: [AdmissionPolicyService]
+  providers: [AdmissionPolicyService],
 })
 export class AdmissionPolicyModule {}
