+++ backend/src/auth/auth.controller.ts
@@ -1,25 +1,13 @@
 import { Body, Controller, Post, Request, UseGuards } from '@nestjs/common';
 import { AuthService } from './auth.service';
 import { LocalAuthGuard } from './guards/local-auth.guards';
-import { User } from '../user/user.entity'; // <-- import your User entity
+import { User } from '../user/user.entity';
 import { CreateUserDto } from '../user/dto/create-user.dto';
-import { JwtAuthGuard } from './guards/jwt-auth.guard';
-import { RolesGuard } from './roles.guard';
-import { Roles } from './roles.decorator';
-import { InviteStudentDto } from './dto/invite-student.dto';
-import { InviteTeacherDto } from './dto/invite-teacher.dto';
-import { CurrentUser } from '../user/current-user.decorator';
-import { InvitationService } from './service/invitation.service';
 import { CredentialLoginDto } from './dto/credential-login.dto';
-import { ActiveUserGuard } from './guards/active-user.guard';
-import { ForgotPinDto } from './dto/forgot-pin.dto';
 
 @Controller('auth')
 export class AuthController {
-  constructor(
-    private authService: AuthService,
-    private invitationService: InvitationService,
-  ) {}
+  constructor(private authService: AuthService) {}
 
   @UseGuards(LocalAuthGuard)
   @Post('login')
@@ -53,55 +41,5 @@ export class AuthController {
       resetPasswordDto.token,
       resetPasswordDto.password,
     );
-  }
-
-  @UseGuards(JwtAuthGuard, ActiveUserGuard, RolesGuard)
-  @Roles('school_admin')
-  @Post('invite/student')
-  inviteStudent(
-    @Body() inviteStudentDto: InviteStudentDto,
-    @CurrentUser() currentUser: User,
-  ) {
-    return this.invitationService.inviteStudent(inviteStudentDto, currentUser);
-  }
-
-  @UseGuards(JwtAuthGuard, ActiveUserGuard, RolesGuard)
-  @Roles('school_admin')
-  @Post('invite/teacher')
-  inviteTeacher(
-    @Body() inviteTeacherDto: InviteTeacherDto,
-    @CurrentUser() currentUser: User,
-  ) {
-    return this.invitationService.inviteTeacher(inviteTeacherDto, currentUser);
-  }
-
-  @UseGuards(JwtAuthGuard, ActiveUserGuard, RolesGuard)
-  @Roles('school_admin')
-  @Post('resend-invitation/student')
-  resendStudentInvitation(
-    @Body() dto: { email: string },
-    @CurrentUser() currentUser: User,
-  ) {
-    return this.invitationService.resendStudentInvitation(
-      dto.email,
-      currentUser,
-    );
-  }
-
-  @UseGuards(JwtAuthGuard, ActiveUserGuard, RolesGuard)
-  @Roles('school_admin')
-  @Post('resend-invitation/teacher')
-  resendTeacherInvitation(
-    @Body() dto: { email: string },
-    @CurrentUser() currentUser: User,
-  ) {
-    return this.invitationService.resendTeacherInvitation(
-      dto.email,
-      currentUser,
-    );
-  }
-  @Post('forgot-pin')
-  forgotPin(@Body() forgotPinDto: ForgotPinDto) {
-    return this.invitationService.forgotPin(forgotPinDto.email);
   }
 }
