--- backend/src/auth/dto/invite-user.dto.ts
@@ -1,17 +1,0 @@
-import { IsEmail, IsNotEmpty, IsUUID, IsOptional } from 'class-validator';
-
-export class InviteUserDto {
-  @IsEmail()
-  @IsNotEmpty()
-  email: string;
-
-  @IsNotEmpty()
-  name: string;
-
-  @IsNotEmpty()
-  role: string;
-
-  @IsUUID()
-  @IsOptional()
-  schoolId?: string;
-}
