+++ backend/src/common/modules/email.module.ts
@@ -1,0 +1,8 @@
+import { Module } from '@nestjs/common';
+import { EmailService } from '../services/email.service';
+
+@Module({
+  providers: [EmailService],
+  exports: [EmailService],
+})
+export class EmailModule {}
