+++ backend/src/invitation/dto/complete-registration.dto.ts
@@ -1,0 +1,12 @@
+import { IsNotEmpty, IsString, MinLength } from 'class-validator';
+
+export class CompleteRegistrationDto {
+  @IsNotEmpty()
+  @IsString()
+  token: string;
+
+  @IsNotEmpty()
+  @IsString()
+  @MinLength(8)
+  password: string;
+}
