+++ backend/src/invitation/dto/forgot-pin.dto.ts
@@ -1,7 +1,7 @@
 import { IsEmail, IsNotEmpty } from 'class-validator';
 
 export class ForgotPinDto {
-  @IsEmail()
   @IsNotEmpty()
+  @IsEmail()
   email: string;
 }
