+++ backend/src/invitation/dto/invite-user.dto.ts
@@ -1,0 +1,19 @@
+import { IsEmail, IsNotEmpty, IsString, IsUUID } from 'class-validator';
+
+export class InviteUserDto {
+  @IsNotEmpty()
+  @IsString()
+  name: string;
+
+  @IsNotEmpty()
+  @IsEmail()
+  email: string;
+
+  @IsNotEmpty()
+  @IsUUID()
+  roleId: string;
+
+  @IsNotEmpty()
+  @IsUUID()
+  schoolId: string;
+}
