+++ backend/src/school/school.controller.ts
@@ -12,20 +12,33 @@ import { SchoolService } from './school.service';
 import { School } from './school.entity';
 import { JwtAuthGuard } from 'src/auth/guards/jwt-auth.guard';
 import { RolesGuard } from 'src/auth/roles.guard';
-import { Roles } from 'src/auth/roles.decorator';
+import { Roles } from 'src/auth/decorators/roles.decorator';
 import { ActiveUserGuard } from 'src/auth/guards/active-user.guard';
+import { CreateSchoolDto } from './dto/create-school.dto';
+import { CurrentUser } from 'src/user/current-user.decorator';
+import { User } from 'src/user/user.entity';
 
 @Controller('schools')
 @UseGuards(JwtAuthGuard, ActiveUserGuard, RolesGuard)
 export class SchoolController {
   constructor(private readonly schoolService: SchoolService) {}
 
+  /**
+   * Create a new school
+   * School admins can create one school and will be associated with it
+   */
   @Post()
-  @Roles('super_admin')
-  create(@Body() schoolData: Partial<School>): Promise<School> {
-    return this.schoolService.create(schoolData);
+  @Roles('school_admin')
+  create(
+    @Body() createSchoolDto: CreateSchoolDto,
+    @CurrentUser() user: User,
+  ): Promise<School> {
+    return this.schoolService.create(createSchoolDto, user);
   }
 
+  /**
+   * Get all schools (super admin only)
+   */
   @Get()
   @Roles('super_admin')
   findAll(): Promise<School[]> {
@@ -33,20 +46,43 @@ export class SchoolController {
   }
 
   @Get(':id')
-  @Roles('super_admin', 'admin')
-  findOne(@Param('id') id: string): Promise<School> {
+  @Roles('super_admin', 'school_admin')
+  findOne(@Param('id') id: string, @CurrentUser() user: User): Promise<School> {
+    if (user.role.name === 'school_admin') {
+      // School admin can only view their own school
+      if (!user.school || user.school.id !== id) {
+        throw new Error('You can only view your own school');
+      }
+    }
     return this.schoolService.findOne(id);
   }
 
+  /**
+   * Get the school associated with the current admin user
+   */
+  @Get('my/school')
+  @Roles('school_admin')
+  findMySchool(@CurrentUser() user: User): Promise<School> {
+    return this.schoolService.findByAdmin(user);
+  }
+
+  /**
+   * Update a school
+   * Super admins can update any school, school admins can only update their own
+   */
   @Put(':id')
-  @Roles('super_admin')
+  @Roles('super_admin', 'school_admin')
   update(
     @Param('id') id: string,
     @Body() schoolData: Partial<School>,
+    @CurrentUser() user: User,
   ): Promise<School> {
-    return this.schoolService.update(id, schoolData);
+    return this.schoolService.update(id, schoolData, user);
   }
 
+  /**
+   * Delete a school (super admin only)
+   */
   @Delete(':id')
   @Roles('super_admin')
   remove(@Param('id') id: string): Promise<void> {
