+++ backend/src/user/user.module.ts
@@ -6,13 +6,11 @@ import { User } from './user.entity';
 import { RoleService } from 'src/role/role.service';
 import { Role } from 'src/role/role.entity';
 import { School } from 'src/school/school.entity';
-import { UserInvitationService } from './user-invitation.service';
-import { UserInvitationController } from './user-invitation.controller';
 
 @Module({
   imports: [TypeOrmModule.forFeature([User, Role, School])],
-  providers: [UserService, RoleService, UserInvitationService],
-  controllers: [UserController, UserInvitationController],
-  exports: [UserService, UserInvitationService],
+  providers: [UserService, RoleService],
+  controllers: [UserController],
+  exports: [UserService],
 })
 export class UserModule {}
